chore: report to sentry if the instance generation has changed (#134)
If spreadReconciling is enabled we know if the generation change of a custom resource has already been observed. If it has been observed we do not report to sentry. This change should reduce the number of unnecessary sentry alerts that occur when a new version of an operator is deployed.

Authored-by: Jonas Lingg <j.lingg@sap.com>